# 18-API/openAPI(Swagger).yml
openapi: 3.0.0
info:
  title: 18-API
  description: 18-API
  version: 1.0.1
servers:
  - url: http://localhost:8000/
    description: Локальный сервер
paths:
  /promotions/create-promotion:
    post:
      summary: создать промоакцию
      requestBody:
        description: создать промоакцию
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPromotion'
      responses:
        '200':
          description: промоакция создана
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Promotion'
  /promotions/update-promotion-status:
    post:
      summary: обновить статус промо-акции
      requestBody:
        description: обновление промо-акцию
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePromotionStatus'
      responses:
        '200':
          description: статус обновлен
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Promotion'
  /promotions/update-promotion:
    post:
      summary: обновить промо-акцию
      requestBody:
        description: обновление статуса промоакции
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePromotion'
      responses:
        '200':
          description: статус обновлен
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Promotion'
  /promotions/delete-promotion:
    delete:
      summary: удалить промоакцию по ID
      requestBody:
        description: удаление промоакции
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeletePromotionRequest'
      responses:
        '200':
          description: акция удалена
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/DeletePromotionResponse'
  /promotions/get-promotion-by-id:
    get:
      summary: получить промоакцию по id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetPromotionByIdRequest'
      responses:
        '200':
          description: промоакция получена
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Promotion'
  /promotions/getpromotions:
    get:
      summary: получить промоакции
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetPromotions'
      responses:
        '200':
          description: промоакции получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Promotion'
  /Items/create-item:
    post:
      summary: создать товар
      requestBody:
        description: создать товар
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewItem'
      responses:
        '200':
          description: товар создан
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Item'
  /Items/update-item-status:
    post:
      summary: обновить статус товара
      requestBody:
        description: обновление статуса товара
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemStatus'
      responses:
        '200':
          description: статус обновлен
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Item'
  /Items/update-item:
    post:
      summary: обновить товар
      requestBody:
        description: обновление товара
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItem'
      responses:
        '200':
          description: статус обновлен
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Item'
  /Items/deleteItem:
    delete:
      summary: удалить товара по ID
      requestBody:
        description: удаление товара
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteItemRequest'
      responses:
        '200':
          description: товар удален
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/DeleteItemResponse'
  /Items/get-item-by-id:
    get:
      summary: получить товар по id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetItemByIdRequest'
      responses:
        '200':
          description: товар получен
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Item'
  /Items/getItems:
    get:
      summary: получить товары
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetItems'
      responses:
        '200':
          description: товары получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
###
components:
  schemas:
    Promotion:
      type: object
      properties:
        id:
          type: number
          description: идентификатор акции
        name:
          type: string
          description: название промоакции
        description:
          type: string
          description: описание
        discoutnPercent:
          type: number
          description: процент скидки
        objectStatusId:
          type: number
          description: цифровой номер статуса записи
        userId":
          type: number
          description: идентификатор пользователя       
    UpdatePromotionStatus:
      properties:
        id:
          type: number
          description: идентификатор акции{
        newStatusId:
          type: number
          description: цифровой номер статуса записи
    NewPromotion:
      type: object
      properties:
        name:
          type: string
          description: название промоакции
        description:
          type: string
          description: описание
        discoutnPercent:
          type: number
          description: процент скидки
        objectStatusId:
          type: number
          description: цифровой номер статуса записи
        userId":
          type: number
          description: идентификатор пользователя
    GetPromotions:
      type: object
      properties:
        getParams:
          type: object
          description: параметры запроса
          properties:
            searchParams:
              type: object
              properties:
                name:
                  type: string
                objectStatusId:
                  type: number
            sortParams:
              type: object
              properties:
                name:
                  type: string
                  enum: ["asc", "desc"]
                objectStatusId:
                  type: string
                  enum: ["asc", "desc"]
    UpdatePromotion:
      type: object
      properties:
        id:
          type: number
          description: идентификатор акции
        name:
          type: string
          description: название промоакции
        description:
          type: string
          description: описание
        discoutnPercent:
          type: number
          description: процент скидки
        objectStatusId:
          type: number
          description: цифровой номер статуса записи
        userId":
          type: number
          description: идентификатор пользователя
    DeletePromotionResponse:
      type: object
      properties:
        message:
          type: string
        id:
          type: number
          enum: [0, 1, 2]
    DeletePromotionRequest:
      type: object
      properties:
        id:
          type: number
    GetPromotionByIdRequest:
      type: object
      properties:
        id:
          type: number
    Item:
      type: object
      description: товар
      properties:
        id:
          type: number
          description: идентификатор товара
        name:
          type: string
          description: название товара
        description:
          type: string
          description: описание товара
        price:
          type: number
          description: цена товара
        userId":
          type: number
          description: идентификатор пользователя
        objectStatusId:
          type: number
          description: цифровой номер статуса записи
        itemTypeId:
          type: number
          description: цифровой номер типа товара
        ItemBalance":
          type: object
          description: складские остатки
          properties:
            id:
              type: number
              description: идентификатор записи
            name:
              type: string
              description: название
            description:
              type: string
              description: описание
            itemId:
              type: number
              description: Идентификатор товара
            userId":
              type: number
              description: идентификатор пользователя
            objectStatusId:
              type: number
              description: цифровой номер статуса записи
            quantity:
              type: number
              description: остаток товара
    UpdateItemStatus:
      properties:
        id:
          type: number
          description: идентификатор товара
        newStatusNumber:
          type: number
          description: цифровой номер статуса записи
    NewItem:
      type: object
      properties:
        name:
          type: string
          description: название товара
        description:
          type: string
          description: описание товара
        price:
          type: number
          description: цена товара
        objectStatusId:
          type: number
          description: цифровой номер статуса записи
        itemTypeId:
          type: number
          description: цифровой номер типа товара
    GetItems:
      type: object
      properties:
        getParams:
          type: object
          description: параметры запроса
          properties:
            name:
              type: string
              description: наименование товара товара
            minPrice:
              type: number
              description: минимальная цена
            maxPrice:
              type: number
              description: максимальная цена
            itemTypeId:
              type: array
              description: массив с кодами типов товара
              items:
                type: number
            objectStatusId:
              type: array
              description: массив с кодами статусов товара
              items:
                type: number
        sortParams:
          type: object
          description: параметры сортировки
          properties:
            name:
              type: string
              description: наименование товара товара
              enum: ["asc", "desc"]
            price:
              type: string
              description: цена товара
              enum: ["asc", "desc"]
            itemTypeId:
              type: string
              description: тип товара
              enum: ["asc", "desc"]
            objectStatusId:
              type: string
              description: статус товара
              enum: ["asc", "desc"]
    UpdateItem:
      type: object
      properties:
        id:
          type: number
          description: идентификатор товара
        name:
          type: string
          description: название товара
        description:
          type: string
          description: описание товара
        price:
          type: number
          description: цена товара
        userId":
          type: number
          description: идентификатор пользователя
        objectStatusId:
          type: number
          description: цифровой номер статуса записи
        itemTypeId:
          type: number
          description: цифровой номер типа товара
    DeleteItemResponse:
      type: object
      properties:
        message:
          type: string
        id:
          type: number
          enum: [0, 1, 2]
    DeleteItemRequest:
      type: object
      properties:
        id:
          type: number
          description: идентификатор товара
    GetItemByIdRequest:
      type: object
      properties:
        id:
          type: number
          description: идентификатор товара